homework-03/src/main/java/ru/barmatin/homework03/service/QuestionServiceImpl.java
package ru.barmatin.homework03.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import ru.barmatin.homework03.dao.QuestionDao;
import ru.barmatin.homework03.domain.Question;
import ru.barmatin.homework03.exception.QuestionsLoadingException;

import java.util.List;

@Service
public class QuestionServiceImpl implements QuestionService {
    private final QuestionDao questionDao;

    @Autowired
    public QuestionServiceImpl(QuestionDao questionDao) {
        this.questionDao = questionDao;
    }

    public List<Question> getQuestionList() throws QuestionsLoadingException {
        return questionDao.getQuestionList();

    }
}
